feat: add image orchestration pipeline with cache, generation, and storage
- Add StepImagePipeline: embed -> cache check -> generate -> GCS upload -> cache store
- Add find_similar_step_image with exact + vector similarity search to DatabaseService
- Add save_step_image_cache and update_step_image_url DB methods
- Local image-gen with fal.ai fallback, progressive Redis publishing
